UI Update
Buttons have been added.
UI now uses pixels for position.
UI can now be "Anchored" to one of nine positions.
UI Scaling now works a lot better.
Moved the input controller event logic from game loop into its class.
Game loop now calls one method for input handling.

#Includes my networking stuff for scott's class

+++ Thomas/ThomasTheGameEngine/ThomasTheGameEngine/GameProperties.cpp
@@ -88,5 +88,5 @@ void GameProperties::VideoProperties::writeValues()
 
 float GameProperties::VideoProperties::aspectRatio()
 {
-	return screenWidth / screenHeight;
+	return screenWidth / (float)screenHeight;
 }
